pyproject.toml
[project]
name = "chess-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polyglot>=16.7.4",
    "pygame>=2.6.1",
    "pygame-ce>=2.5.5",
    "python-chess>=1.999",
]
